fix: ticket agent scroll
