+++ desk/src/pages/SearchAgent.vue
@@ -1,15 +1,13 @@
 <template>
   <div>
     <div>
-      <header
-        class="sticky top-0 z-10 border-b bg-surface-white px-4 py-2.5 sm:px-5"
-      >
-        <div class="flex items-center justify-between">
+      <LayoutHeader>
+        <template #left-header>
           <Breadcrumbs
             :items="[{ label: 'Search', route: { name: 'SearchAgent' } }]"
           />
-        </div>
-      </header>
+        </template>
+      </LayoutHeader>
       <div class="mx-auto mt-6 max-w-4xl px-4 sm:px-5">
         <div class="flex items-center space-x-2 px-2.5">
           <TextInput
@@ -195,28 +193,25 @@
 </template>
 
 <script setup lang="ts">
-import { ref, onMounted, computed, useTemplateRef, watch } from "vue";
-import { useRouter, useRoute } from "vue-router";
+import SearchMultiSelect from "@/components/SearchMultiSelect.vue";
 import {
   Breadcrumbs,
-  TextInput,
+  createResource,
   debounce,
-  Tooltip,
   ErrorMessage,
+  TextInput,
 } from "frappe-ui";
-import { createResource } from "frappe-ui";
-import UserAvatar from "@/components/UserAvatar.vue";
-import SearchMultiSelect from "@/components/SearchMultiSelect.vue";
+import { computed, onMounted, ref, useTemplateRef, watch } from "vue";
+import { useRoute, useRouter } from "vue-router";
 // import { vFocus } from "@/directives";
 
 // Icons
-import LucideSearch from "~icons/lucide/search";
-import LucideX from "~icons/lucide/x";
-import LucideThumbsUp from "~icons/lucide/thumbs-up";
-import LucideThumbsDown from "~icons/lucide/thumbs-down";
-import LucideTicket from "~icons/lucide/ticket";
-import LucideMessageSquare from "~icons/lucide/message-square";
+import { LayoutHeader } from "@/components";
 import LucideMail from "~icons/lucide/mail";
+import LucideMessageSquare from "~icons/lucide/message-square";
+import LucideSearch from "~icons/lucide/search";
+import LucideTicket from "~icons/lucide/ticket";
+import LucideX from "~icons/lucide/x";
 
 // Type Definitions
 interface SearchSummary {
